requirements.txt
Django==5.2.10
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
django-csp==4.0
whitenoise==6.7.0
gunicorn==22.0.0
dj-database-url==2.2.0
PyJWT==2.10.1
asgiref==3.11.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
inflection==0.5.1
PyYAML==6.0.3
jsonschema==4.26.0
attrs==25.4.0
referencing==0.37.0
rpds-py==0.30.0
jsonschema-specifications==2025.9.1
typing-extensions==4.15.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
openpyxl==3.1.2
reportlab
requests==2.31.0

# Local dev helpers (optional)
django-extensions
Werkzeug
pyOpenSSL
cryptography
